Corregido tipo de nroCliente

+++ src/main/java/com/sistema/application/models/ClienteModel.java
@@ -9,11 +9,11 @@ import org.springframework.format.annotation.NumberFormat.Style;
 public class ClienteModel extends PersonaModel {
 	private String email;
 	@NumberFormat(style = Style.NUMBER)
-	private int nroCliente;
+	private long nroCliente;
 	
 	public ClienteModel() {}
 	
-	public ClienteModel(long id, String nombre, String apellido, int dni, LocalDate fechaNacimiento, String email, int nroCliente) {
+	public ClienteModel(long id, String nombre, String apellido, int dni, LocalDate fechaNacimiento, String email, long nroCliente) {
 		super(id, nombre, apellido, dni, fechaNacimiento);
 		this.email = email;
 		this.nroCliente = nroCliente;
@@ -37,10 +37,10 @@ public class ClienteModel extends PersonaModel {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public int getNroCliente() {
+	public long getNroCliente() {
 		return nroCliente;
 	}
-	public void setNroCliente(int nroCliente) {
+	public void setNroCliente(long nroCliente) {
 		this.nroCliente = nroCliente;
 	}
